+++ grid-proxy/tests/queries/node_test.go
@@ -276,6 +276,10 @@ var nodeFilterRandomValueGenerator = map[string]func(agg NodesAggregate) interfa
 		v := uint64(rand.Intn(1100)) // 1000 is the total nodes + 100 for non-existed cases
 		return &v
 	},
+	"NodeIDs": func(agg NodesAggregate) interface{} {
+		randomLen := rand.Intn(5)
+		return getRandomSliceFrom(agg.nodeIDs, randomLen)
+	},
 	"TwinID": func(agg NodesAggregate) interface{} {
 		v := uint64(rand.Intn(3500))
 		return &v